requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
colorama==0.4.6
python-dotenv==1.0.0
anthropic==0.19.1